integracion redis

+++ internal/domain/user/user.go
@@ -3,7 +3,7 @@ package user
 import "time"
 
 type User struct {
-	ID           int64      `json:"id"`
+	ID           int        `json:"id"`
 	Username     string     `json:"username"`
 	Email        string     `json:"email"`
 	PasswordHash string     `json:"-"` // No se expone en JSON
@@ -13,7 +13,7 @@ type User struct {
 	BirthDate    *time.Time `json:"birth_date,omitempty"`
 	IsActive     bool       `json:"is_active"`
 
-	CountryID   int32   `json:"country_id"`
+	CountryID   int     `json:"country_id"`
 	AddressLine *string `json:"address_line,omitempty"`
 
 	CreatedAt time.Time  `json:"created_at"`
